first version of gui integration with known issues

+++ Ants/src/com/ibm/sk/dto/Warrior.java
@@ -7,7 +7,7 @@ import com.ibm.sk.dto.enums.Direction;
 
 public class Warrior extends AbstractWarrior {
 
-	public Warrior(final int id, final Point position, final Hill myHill) {
+	public Warrior(final long id, final Point position, final Hill myHill) {
 		super(id, position, myHill);
 	}
